+++ python_modules/dagster/dagster/cli/schedule.py
@@ -19,6 +19,7 @@ def create_schedule_cli_group():
     group.add_command(schedule_preview_command)
     group.add_command(schedule_start_command)
     group.add_command(schedule_stop_command)
+    group.add_command(schedule_restart_command)
     return group
 
 
@@ -260,7 +261,7 @@ def execute_start_command(schedule_name, all_flag, cli_args, print_fn):
             try:
                 schedule = scheduler.start_schedule(schedule.name)
             except DagsterInvariantViolationError as ex:
-                continue
+                raise click.UsageError(ex)
 
         print_fn("Started all schedules for repository {name}".format(name=repository.name))
     else:
@@ -307,3 +308,63 @@ def execute_stop_command(schedule_name, cli_args, print_fn):
             schedule_name=schedule_name, schedule_id=schedule.schedule_id
         )
     )
+
+
+@click.command(name='restart', help="Restart a running schedule")
+@click.argument('schedule_name', nargs=-1)
+@click.option(
+    '--restart-all-running',
+    help="restart previously running schedules",
+    is_flag=True,
+    default=False,
+)
+@repository_target_argument
+def schedule_restart_command(schedule_name, restart_all_running, **kwargs):
+    schedule_name = extract_schedule_name(schedule_name)
+    return execute_restart_command(schedule_name, restart_all_running, kwargs, click.echo)
+
+
+def execute_restart_command(schedule_name, all_running_flag, cli_args, print_fn):
+    handle = handle_for_repo_cli_args(cli_args)
+    repository = handle.build_repository_definition()
+
+    instance = DagsterInstance.get()
+    schedule_handle = handle.build_scheduler_handle(artifacts_dir=instance.schedules_directory())
+
+    if not schedule_handle:
+        print_fn("Scheduler not defined for repository {name}".format(name=repository.name))
+        return
+
+    scheduler = schedule_handle.get_scheduler()
+    if all_running_flag:
+        for schedule in scheduler.all_schedules():
+            if schedule.status == ScheduleStatus.RUNNING:
+                try:
+                    scheduler.stop_schedule(schedule.name)
+                    scheduler.start_schedule(schedule.name)
+                except DagsterInvariantViolationError as ex:
+                    raise click.UsageError(ex)
+
+        print_fn(
+            "Restarted all running schedules for repository {name}".format(name=repository.name)
+        )
+    else:
+        schedule = scheduler.get_schedule_by_name(schedule_name)
+        if schedule.status != ScheduleStatus.RUNNING:
+            click.UsageError(
+                "Cannot restart a schedule {name} because is not currently running".format(
+                    name=schedule.name
+                )
+            )
+
+        try:
+            scheduler.stop_schedule(schedule_name)
+            scheduler.start_schedule(schedule_name)
+        except DagsterInvariantViolationError as ex:
+            raise click.UsageError(ex)
+
+        print_fn(
+            "Restarted schedule {schedule_name} with ID {schedule_id}".format(
+                schedule_name=schedule_name, schedule_id=schedule.schedule_id
+            )
+        )